separate nightly tasks and other scheduled tasks.
other tasks is anything that is run on a different frequency than
nightly
