+++ tests/app/celery/test_tasks.py
@@ -15,7 +15,6 @@ from notifications_utils.columns import Row
 from app import (encryption, DATETIME_FORMAT)
 from app.celery import provider_tasks
 from app.celery import tasks
-from app.celery.scheduled_tasks import check_job_status
 from app.celery.tasks import (
     process_job,
     process_row,
@@ -1394,24 +1393,6 @@ def test_send_inbound_sms_to_service_does_not_retries_if_request_returns_404(not
         send_inbound_sms_to_service(inbound_sms.id, inbound_sms.service_id)
 
     mocked.call_count == 0
-
-
-def test_check_job_status_task_does_not_raise_error(sample_template):
-    create_job(
-        template=sample_template,
-        notification_count=3,
-        created_at=datetime.utcnow() - timedelta(hours=2),
-        scheduled_for=datetime.utcnow() - timedelta(minutes=31),
-        processing_started=datetime.utcnow() - timedelta(minutes=31),
-        job_status=JOB_STATUS_FINISHED)
-    create_job(
-        template=sample_template,
-        notification_count=3,
-        created_at=datetime.utcnow() - timedelta(minutes=31),
-        processing_started=datetime.utcnow() - timedelta(minutes=31),
-        job_status=JOB_STATUS_FINISHED)
-
-    check_job_status()
 
 
 def test_process_incomplete_job_sms(mocker, sample_template):
